# river: have `concat` bypass decoder