--- a/pkg/river/internal/value/value.go
+++ b/pkg/river/internal/value/value.go
@@ -24,6 +24,7 @@ var (
 	goDuration        = reflect.TypeOf((time.Duration)(0))
 	goDurationPtr     = reflect.TypeOf((*time.Duration)(nil))
 	goRiverDecoder    = reflect.TypeOf((*Unmarshaler)(nil)).Elem()
+	goRawRiverFunc    = reflect.TypeOf((RawFunction)(nil))
 )
 
 // NOTE(rfratto): This package is extremely sensitive to performance, so
@@ -258,6 +259,9 @@ func (v Value) Interface() interface{} {
 	return v.rv.Interface()
 }
 
+// Reflect returns the raw reflection value backing v.
+func (v Value) Reflect() reflect.Value { return v.rv }
+
 // makeValue converts a reflect value into a Value, deferencing any pointers or
 // interface{} values.
 func makeValue(v reflect.Value) Value {
@@ -390,6 +394,10 @@ func (v Value) Key(key string) (index Value, ok bool) {
 func (v Value) Call(args ...Value) (Value, error) {
 	if v.ty != TypeFunction {
 		panic("river/value: Call called on non-function type")
+	}
+
+	if v.rv.Type() == goRawRiverFunc {
+		return v.rv.Interface().(RawFunction)(v, args...)
 	}
 
 	var (